+++ logprep/generator/http/controller.py
@@ -3,6 +3,7 @@ This generator will parse example events, manipulate their timestamps and send t
 a defined output
 """
 
+import json
 import logging
 import time
 from concurrent.futures import ThreadPoolExecutor
@@ -24,6 +25,7 @@ class Controller:
 
     def __init__(self, **kwargs):
         self.config = kwargs
+        self.loghandler = None
         self._setup_logging()
         self.thread_count: int = kwargs.get("thread_count")
         self.use_reporter: bool = kwargs.get("report")
@@ -40,16 +42,17 @@ class Controller:
 
     def _setup_logging(self):
         console_logger = logging.getLogger("console")
+        if self.config.get("loglevel"):
+            logging.getLogger().setLevel(self.config.get("loglevel"))
         if console_logger.handlers:
             console_handler = console_logger.handlers.pop()  # last handler is console
             self.loghandler = LogprepMPQueueListener(logqueue, console_handler)
             self.loghandler.start()
 
-    def run(self):
+    def run(self) -> str:
         """
         Iterate over all event classes, trigger their processing and count the return statistics
         """
-        # TODO use a logprep pipeline to handle processing
         logger.info("Started Data Processing")
         self.input.reformat_dataset()
         run_time_start = time.perf_counter()
@@ -59,10 +62,39 @@ class Controller:
             logger.info("Gracefully shutting down...")
         self.input.clean_up_tempdir()
         run_duration = time.perf_counter() - run_time_start
-        stats = self.output.metrics.status_codes.tracker.collect()[0].values
-        logger.info("Completed with following http return code statistics: %s", stats)
+        stats = self._get_statistics()
+        logger.info("Completed with following statistics: %s", stats)
         logger.info("Execution time: %f seconds", run_duration)
-        self.loghandler.stop()
+        if self.loghandler is not None:
+            self.loghandler.stop()
+        return stats
+
+    def _get_statistics(self) -> str:
+        status_code_stats = self.output.metrics.status_codes.tracker.collect()[0].samples
+        status_code_stats = {
+            f'Requests http status {sample.labels.get("description")}': sample.value
+            for sample in status_code_stats
+            if sample.name.endswith("total")
+        }
+        sucessfull_events_stats = self.output.metrics.number_of_processed_events.tracker.collect()[
+            0
+        ].samples
+        sucessfull_events_stats = {
+            "Events successfull": sample.value
+            for sample in sucessfull_events_stats
+            if sample.name.endswith("total")
+        }
+        failed_events_stats = self.output.metrics.number_of_failed_events.tracker.collect()[
+            0
+        ].samples
+        failed_events_stats = {
+            "Events failed": sample.value
+            for sample in failed_events_stats
+            if sample.name.endswith("total")
+        }
+        stats = sucessfull_events_stats | status_code_stats | failed_events_stats
+        stats = json.dumps(stats, sort_keys=True, indent=4, separators=(",", ": "))
+        return stats
 
     def _generate_load(self):
         with ThreadPoolExecutor(max_workers=self.thread_count) as executor: